Add a 'poop' exchange
The 'poop' emoticon mentioned in the Legend wasn't actually being used
by any row of the table. This fixes it. Verse is a bitcoin exchange
that is part of the 'bitcoin.com' conglomerate, which is an industry
attack on bitcoin because they are infringing bitcoin trademarks by
purporting BCash as the real bitcoin. Therefore, one can easily assume
that they won't ever support Lightning (BCash supporters are always
bashing Lightning, spreading lies about it, such as saying that it's
a custodial technology, or just IOUs, etc). For the record, BCash
doesn't support LN (it can't, as it doesn't have SegWit, therefore
it is prone to transaction malleability).

More info about Verse:
https://cryptoslate.com/nobody-seems-to-trust-roger-vers-newly-announced-cryptocurrency-exchange/