Distinguish between N/A and Unknown
In the majority of cases, the cells marked with n.a didn't really
mean "not applicable", but rather "unknown, please contribute the
info", so I'm marking the latter with a question mark and adding a
legend entry about it. Hopefully thanks to this, we get more
contributions and the table gets more complete.